requirements.txt
# Installed into .venv/ by step-0 (env check)

PyMuPDF>=1.23.0
pdfplumber>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
numpy>=1.24.0
PyYAML>=6.0
rembg>=2.0.0
openai>=1.0.0
edge-tts>=6.0.0
